refactor(core): lift onCopy and onPaste into Provider
