+++ packages/sanity/src/core/form/field/actions/copyAction.ts
@@ -4,7 +4,7 @@ import {useCallback} from 'react'
 
 import {defineDocumentFieldAction} from '../../../config/document/fieldActions/define'
 import {useTranslation} from '../../../i18n'
-import {useCopyPasteAction} from '../../../studio/copyPaste/useCopyPasteAction'
+import {useCopyPaste} from '../../../studio'
 import {useGetFormValue} from '../../contexts/GetFormValue'
 import {type FormDocumentValue} from '../../types/formDocumentValue'
 import {defineActionItem} from './define'
@@ -24,7 +24,7 @@ export const copyAction = defineDocumentFieldAction({
   name: 'copyField',
   useAction({path}) {
     const getFormValue = useGetFormValue()
-    const {onCopy} = useCopyPasteAction()
+    const {onCopy} = useCopyPaste()
     const {t} = useTranslation('copy-paste')
 
     const isDocument = path.length === 0
