+++ packages/sanity/src/core/studio/copyPaste/index.ts
@@ -1,3 +1,2 @@
 export * from './CopyPasteProvider'
 export * from './types'
-export * from './useCopyPasteAction'
